refactor : 내부 DTO 클래스 추출 (#301)
